Refactor savePage a bit to use it as a service from the jobs
